requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
munch==4.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.8
pretrainedmodels==0.7.4
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segmentation_models_pytorch==0.4.0
setuptools==70.2.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
